Refactor source locations for CAS(SCF) utilities
