+++ include/macis/mcscf/diis.hpp
