+++ include/macis/mcscf/mcscf_impl.hpp
@@ -7,15 +7,16 @@
  */
 
 #pragma once
-#include <macis/util/diis.hpp>
 #include <macis/util/fcidump.hpp>
-#include <macis/util/fock_matrices.hpp>
-#include <macis/util/mcscf.hpp>
-#include <macis/util/orbital_gradient.hpp>
-#include <macis/util/orbital_hessian.hpp>
-#include <macis/util/orbital_rotation_utilities.hpp>
-#include <macis/util/orbital_steps.hpp>
 #include <macis/util/transform.hpp>
+
+#include <macis/mcscf/diis.hpp>
+#include <macis/mcscf/fock_matrices.hpp>
+#include <macis/mcscf/mcscf.hpp>
+#include <macis/mcscf/orbital_gradient.hpp>
+#include <macis/mcscf/orbital_hessian.hpp>
+#include <macis/mcscf/orbital_rotation_utilities.hpp>
+#include <macis/mcscf/orbital_steps.hpp>
 
 // #include <ceres/ceres.h>
