+++ src/macis/mcscf/orbital_energies.cxx
@@ -6,7 +6,7 @@
  * See LICENSE.txt for details
  */
 
-#include <macis/util/orbital_energies.hpp>
+#include <macis/mcscf/orbital_energies.hpp>
 
 namespace macis {
