+++ src/macis/moller_plesset.cxx
@@ -9,7 +9,7 @@
 #include <iostream>
 #include <lapack.hh>
 #include <macis/util/moller_plesset.hpp>
-#include <macis/util/orbital_energies.hpp>
+#include <macis/mcscf/orbital_energies.hpp>
 
 namespace macis {
