+++ tests/fock_matrices.cxx
@@ -9,7 +9,7 @@
 #include <iomanip>
 #include <macis/util/detail/rdm_files.hpp>
 #include <macis/util/fcidump.hpp>
-#include <macis/util/fock_matrices.hpp>
+#include <macis/mcscf/fock_matrices.hpp>
 
 #include "ut_common.hpp"
